reorganize benchmarks: sequential & bulk-loading
