+++ bench/Cargo.toml
@@ -15,5 +15,5 @@ criterion = "0.3"
 rand_chacha = "0.3.1"
 
 [[bench]]
-name = "srtree"
+name = "main"
 harness = false
