+++ bench/benches/uniform/mod.rs
@@ -1,0 +1,10 @@
+mod bulkload;
+mod sequential;
+mod utils;
+
+pub use sequential::{
+    sequential_benchmark as uniform_sequential_benchmark,
+};
+pub use bulkload::{
+    bulkload_benchmark as uniform_bulkload_benchmark,
+};
